del fluid include
